+++ src/main/kotlin/lotto/Application.kt
@@ -3,8 +3,6 @@ package lotto
 import lotto.controller.LottoController
 import lotto.presentation.InputViewImpl
 import lotto.presentation.OutputViewImpl
-import lotto.utils.InputValidatorImpl
-import lotto.utils.LottoTicketGeneratorImpl
 
 fun main() {
     val lottoController = LottoController(
